+++ vwo/helpers/impression_util.py
@@ -32,12 +32,13 @@ except ImportError:
 
 FILE = FileNameEnum.Helpers.ImpressionUtil
 
-
 def create_impression(
-    settings_file,
+    vwo_instance,
     campaign_id,
     variation_id,
     user_id,
+    event_properties=None,
+    goal=None,
     goal_id=None,
     revenue=None,
     visitor_user_agent=None,
@@ -67,7 +68,7 @@ def create_impression(
     if goal_id is not None:
         is_track_user_api = False
 
-    impression = get_common_properties(user_id, settings_file, visitor_user_agent, visitor_ip)
+    impression = get_common_properties(user_id, vwo_instance.settings_file, visitor_user_agent, visitor_ip)
 
     impression.update(experiment_id=campaign_id, combination=variation_id)
 
@@ -89,6 +90,8 @@ def create_impression(
         impression.update(goal_id=goal_id)
         if revenue:
             impression.update(r=revenue)
+        elif(vwo_instance.is_event_arch_enabled and event_properties is not None and goal.get('revenueProp') in event_properties ):
+            impression.update(r=event_properties[goal.get('revenueProp')])
         logger.log(
             LogLevelEnum.DEBUG,
             LogMessageEnum.DEBUG_MESSAGES.IMPRESSION_FOR_TRACK_GOAL.format(
@@ -197,7 +200,7 @@ def create_track_user_events_impression(settings_file, campaign_id, variation_id
 
 
 def create_track_goal_events_impression(
-    settings_file, user_id, goal_identifier, campaign_goal_revenue_prop_list, revenue=None
+    settings_file, user_id, goal_identifier, campaign_goal_revenue_prop_list, event_properties=None, revenue=None
 ):
     """Creates the event impression for track goal call from the arguments passed accordingly
 
@@ -233,6 +236,10 @@ def create_track_goal_events_impression(
             campaign_ids=[campaign_id for campaign_id, _, _ in campaign_goal_revenue_prop_list],
         ),
     )
+    if event_properties is not None and len(event_properties) > 0:
+        for prop in event_properties:
+            impression["d"]["event"]["props"].update({prop:event_properties[prop]})
+
     return impression