+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.55.0] - 2023-07-14
+
+### Changed
+
+- Modified the revenueProp logic to read from eventProperties for Data360 enabled accounts
+- Updated the code to check for existence of eventProperties and access the revenueProp property accordingly
+
+## [1.50.1] - 2023-06-26
+
+### Changed
+
+- Changed the test cases for MEG new implementation
+
+## [1.50.0] - 2023-06-26
+
+### Changed
+
+- Update bucketing logic so that variation remains unchanged even after changing campaign traffic distribution. Note: it will not cater the case when user was not part of campaign earlier but started becoming part of campaign after changing the campaign traffic distribution and vice versa.
+
 ## [1.46.1] - 2023-06-08
 
 ### Changed